requirements.txt
mysql-connector-python
uvicorn
fastapi
sentence-transformers
faiss-cpu
google-genai
python-dotenv
pandas
